[Bug] [lang] Fix @ti.func not capturing closure variables
