+++ tests/python/test_syntax_errors.py
@@ -192,3 +192,51 @@ def test_static_grouped_struct_for():
             pass
 
     test()
+
+
+@ti.must_throw(ti.TaichiSyntaxError)
+def test_is():
+    b = ti.var(ti.i32, shape=())
+    c = ti.var(ti.i32, shape=())
+
+    @ti.kernel
+    def func():
+        a = b is c
+
+    func()
+
+
+@ti.must_throw(ti.TaichiSyntaxError)
+def test_is_not():
+    b = ti.var(ti.i32, shape=())
+    c = ti.var(ti.i32, shape=())
+
+    @ti.kernel
+    def func():
+        a = b is not c
+
+    func()
+
+
+@ti.must_throw(ti.TaichiSyntaxError)
+def test_in():
+    b = ti.var(ti.i32, shape=())
+    c = ti.var(ti.i32, shape=())
+
+    @ti.kernel
+    def func():
+        a = b in c
+
+    func()
+
+
+@ti.must_throw(ti.TaichiSyntaxError)
+def test_not_in():
+    b = ti.var(ti.i32, shape=())
+    c = ti.var(ti.i32, shape=())
+
+    @ti.kernel
+    def func():
+        a = b not in c
+
+    func()
